requirements.txt
click==8.0.4
ddddocr==1.4.3
Flask==2.0.3
Flask-Cors==3.0.10
flatbuffers==2.0
importlib-metadata==4.11.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
numpy==1.21.5
onnxruntime==1.10.0
opencv-python==4.5.5.64
Pillow==9.0.1
protobuf==3.19.4
six==1.16.0
typing_extensions==4.1.1
uWSGI==2.0.20
Werkzeug==2.0.3
zipp==3.7.0
